Clean Code + Modify + Seeder + Fixing

+++ database/migrations/2024_08_08_035114_create_pembelian_barang_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
             $table->string('id_users')->nullable();
             $table->string('no_nota')->nullable();
             $table->dateTime('tgl_nota')->nullable();
-            $table->double('total_item')->nullable();
-            $table->double('total_nilai')->nullable();
+            $table->integer('total_item')->nullable();
+            $table->decimal('total_nilai', 15, 2)->nullable();
             $table->enum('status', ['progress', 'success', 'failed','mixed'])->default('progress');
             $table->softDeletes();
         });
