+++ database/migrations/2024_08_13_062114_create_detail_pembelian_barang_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->string('id_pembelian_barang')->nullable();
             $table->string('id_barang')->nullable();
             $table->integer('qty')->nullable();
-            $table->double('harga_barang')->nullable();
-            $table->double('total_harga')->nullable();
+            $table->decimal('harga_barang', 15, 2)->nullable();
+            $table->decimal('total_harga', 15, 2)->nullable();
             $table->enum('status', ['success', 'failed', 'progress'])->default('progress')->nullable();
             $table->timestamps();
             $table->softDeletes();
