+++ database/migrations/2024_08_26_082913_create_pengiriman_barang_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('ekspedisi');
             $table->string('toko_penerima');
             $table->dateTime('tgl_kirim');
-            $table->double('total_item')->nullable();
-            $table->double('total_nilai')->nullable();
+            $table->integer('total_item')->nullable();
+            $table->decimal('total_nilai', 15, 2)->nullable();
             $table->dateTime('tgl_terima')->nullable();
             $table->enum('status', ['pending', 'progress', 'success', 'canceled'])->default('pending');
             $table->enum('tipe_pengiriman', ['mutasi', 'reture'])->default('mutasi');
